+++ aom/aom.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 
 	"github.com/philippgille/apiomat-go/aomc"
-	"github.com/philippgille/apiomat-go/aomm"
+	"github.com/philippgille/apiomat-go/aoms"
 )
 
 func main() {
@@ -17,15 +17,15 @@ func main() {
 	flag.Parse()
 
 	// Version
-	aommClient := aomm.NewAomClient(*baseUrl, "", "", "")
-	version, err := aommClient.GetVersion()
+	aomsClient := aoms.NewAomClient(*baseUrl, "", "", "")
+	version, err := aomsClient.GetVersion()
 	if err != nil {
 		panic(err)
 	}
 	fmt.Println(version) // {"server":"null:443","version":"2.6.2-107E"}
 
 	// Classes
-	aomcClient := aomc.NewAomClient(aommClient.BaseUrl, *username, *password, "")
+	aomcClient := aomc.NewAomClient(aomsClient.BaseUrl, *username, *password, "")
 	classes, err := aomcClient.GetClasses(*module, "")
 	if err != nil {
 		panic(err)