+++ RELEASES.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 vNext
 -----
 
+- Changed: Renamed package aomm to aoms
+
 v0.1.0 (2018-04-15)
 -------------------